Struct automerge_backend::Backend

pub struct Backend { /* fields omitted */ }

Implementations

impl Backend

pub fn new() -> Self

pub fn load_changes(
&mut self,
changes: Vec<Change>
) -> Result<(), AutomergeError>

pub fn apply_changes(
&mut self,
changes: Vec<Change>
) -> Result<Patch, AutomergeError>

pub fn get_heads(&self) -> Vec<ChangeHash>

pub fn apply_local_change(
&mut self,
change: Change
) -> Result<(Patch, Change), AutomergeError>

pub fn get_patch(&self) -> Result<Patch, AutomergeError>

pub fn get_changes_for_actor_id(
&self,
actor_id: &ActorId
) -> Result<Vec<&Change>, AutomergeError>

pub fn get_changes(&self, have_deps: &[ChangeHash]) -> Vec<&Change>

pub fn save(&self) -> Result<Vec<u8>, AutomergeError>

pub fn load(data: Vec<u8>) -> Result<Self, AutomergeError>

pub fn get_missing_deps(&self, heads: &[ChangeHash]) -> Vec<ChangeHash>

pub fn get_change_by_hash(&self, hash: &ChangeHash) -> Option<&Change>

pub fn get_changes_added<'a>(&self, other: &'a Self) -> Vec<&'a Change>

Returns all changes that are present in self but not present in other.

pub fn add_event_handler(&mut self, handler: EventHandler) -> EventHandlerId

Adds the event handler and returns the id of the handler.

pub fn remove_event_handler(&mut self, id: EventHandlerId) -> bool

Remove the handler with the given id, returning whether it removed a handler or not.

impl Backend

pub fn generate_sync_message(
&self,
sync_state: &mut SyncState
) -> Option<SyncMessage>

pub fn receive_sync_message(
&mut self,
sync_state: &mut SyncState,
message: SyncMessage
) -> Result<Option<Patch>, AutomergeError>

pub fn get_changes_to_send(
&self,
have: Vec<SyncHave>,
need: &[ChangeHash]
) -> Vec<&Change>